requirements.txt
alembic==1.17.2
apache-airflow==3.1.3
argparse==1.4.0
docker==7.1.0
fastapi==0.117.1
greenlet==3.2.4
joblib==1.5.2
mlflow==3.6.0
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.4
pytest==9.0.1
PyYAML==6.0.3
scikit-learn==1.7.2
SQLAlchemy==2.0.44
typing-inspection==0.4.2
uvicorn==0.38.0
xgboost==3.1.2
yfinance==0.2.66